#16 Design Home Page
The most significant changes in the code are related to the enhancement of the `HomePage.xaml` and `HomeViewModel.cs` files, the modification of the `Amount` property in the `Transactions` table, and the addition of a new migration file.

In `HomePage.xaml`, a new `ToolbarItem` named "Quick Add" has been added to the `ContentPage.ToolbarItems` collection. The structure of the `ScrollView` and `Grid` elements has been changed, with new `Frame` elements added for displaying account, loan, and auto-pay information. The "Quick Add" frame has been expanded to include fields for adding transactions, accounts, and loans. The `ScrollView` has been given a name, "HomeScroll", for reference in the code-behind file.

In `HomeViewModel.cs`, several new properties have been added to the `HomeViewModel` class, including collections for accounts, loans, and auto-pays, selected items for each of these collections, visibility flags for the chart and quick add form, and headers for the account, loan, and auto-pay sections. The `AddTransaction`, `AddAccount`, and `AddLoan` methods have been updated to create and add new instances of `TransactionModel`, `AccountModel`, and `LoanModel` respectively. A new `QuickAdd` method has been added to toggle the visibility of the quick add form and send a message to scroll to it when it becomes visible.

In `ApplicationDbContextModelSnapshot.cs` and `TransactionModel.cs`, the `Amount` property of the `Transactions` table has been changed from `decimal` to `double`.

A new migration `20240521034120_ChangesForIssue16.cs` has been added to update the `Amount` column in the `Transactions` table from `decimal` to `double`. This migration file is auto-generated and is used to apply changes to the database schema. The changes include the creation of several tables and their relationships. The tables created include: `UserBuckets`, `FinanceAccounts`, `AutoPays`, `SavingsBuckets`, `Loans`, `Settings`, and `Transactions`.

Other minor changes include the removal of the default value of `Id` property in `AccountViewModel.cs`, the initialization of the `LoanName` property to an empty string in `LoanModel.cs`, and the registration of two new messages with the `WeakReferenceMessenger` in `HomePage.xaml.cs`.

+++ MoneyTrackerMigrations/Models/TransactionModel.cs
@@ -9,7 +9,7 @@ public class TransactionModel
     [Required]
     public string Name { get; set; } = string.Empty;
     [Required]
-    public decimal Amount { get; set; }
+    public double Amount { get; set; }
     public string? Notes { get; set; }
     [Required]
     public DateTime Date { get; set; } = DateTime.Now.Date;
